requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
Flask-Login~=0.6.3
WTForms~=3.2.1
python-dotenv~=1.0.1